fix build against system 0.14.0 library on Debian unstable
In the Debian package, we build without the httplib submodule. This
somehow triggers this code path which fails with:

    --   Package dependency requirement 'cpp-httplib <= 0.14' could not be satisfied.
    Package 'cpp-httplib' has version '0.14.0', required version is '<= 0.14'

This rather counter-intuitive message actually says that 0.14.0, the
version found in pkgconfig is *higher* than 0.14. That seems bizarre,
but it's the way this works.

Since we still want to restrict the upper bound because httplib
tends to break the API between releases, let's use the *next* release
as a target.

An alternative would have been to simply not specify an upper bound or
use the minor version as well (e.g. `<=0.14.0`) but that seems rather
more brittle.