Fix main window geometry restoration
If we call restoreGeometry() after main loop iteration,
it might happen in the middle of creation of native window and be ignored.
So call synchronously in constructor instead.

The reason why we were doing it like that in the first place
is because in Impl constructor the MainWindow::mImpl member is still uninitialized,
and we use in MainWindow::event override, which may be called from restoreGeometry().
Instead use QMetaObject::invokeMethod in that override so that Impl::updateShowHideAction
is called on the next iterator of event loop (this is what we should be doing anyway).