Remove checkSingleFile request and instead just check the existence of incomplete file
This request was needed to know whether the torrent is single file or a directory with many files.
The reason why we needed to know this is because renaming of incomplete files works only on files, not directories,
so when user clicks on "Open" menu item, we need to add .part to the path only if the torrent is a file.
However this approach was flawed: it incorrectly handled the case of a torrent with a single file inside directory.

Instead we can simply check if a file with .part suffix exists, and open it only in that case.